requirements.txt
Pillow
matplotlib
numpy
seaborn
torch
torchvision
tqdm
